Add spatial mysql2 adapter functionality

+++ db/migrate/20150427002952_recreate_notes_with_spatial_index.rb
@@ -1,21 +1,16 @@
 class RecreateNotesWithSpatialIndex < ActiveRecord::Migration
   def up
-    drop_table :notes
+    drop_table :notes if table_exists?(:notes)
 
     create_table :notes, :options => 'ENGINE=MyISAM' do |t|
       t.string :title
       t.string :body
-      t.decimal :latitude, precision: 9, scale: 6
-      t.decimal :longitude, precision: 9, scale: 6
-
+      t.column :latlng, :point, null: false
       t.timestamps null: false
     end
 
+    add_index :notes, :latlng, spatial: true
     add_reference :notes, :user, index: true, foreign_key: true
-
-    # Add spatial indexing on a point (lat, lng)
-    execute 'ALTER TABLE notes ADD pt POINT NOT NULL;'
-    execute 'ALTER TABLE notes ADD SPATIAL INDEX (pt);'
   end
 
   def down
